Update code.ipynb
Added some identifiers to overfitting. However, value dropped actually lower the precision of random forest classifier and did not really help with other classifiers.

Note: Code blocks on KNN runs for very long >1h.

really think that the logistic regression and knn models are just not good for this dataset. cant find a way to make those better. May be able to find a way to better deal with the overfitting.